+++ src/components/admin/assets/AssetsTable.vue
@@ -9,9 +9,7 @@
       :routeEditName='routeEditName',
       :storeGetter='storeGetter',
       :storeRemove='storeRemove',
-      :subtitle='subtitle',
       :tableCols='tableCols',
-      :title='title',
       )
 </template>
 
@@ -37,7 +35,6 @@ export default {
         { label: 'Name', prop: 'name' },
         { label: 'Type', prop: 'type' },
       ],
-      title: 'Assets',
     }
   },
 }