+++ src/components/admin/languages/LanguagesTable.vue
@@ -9,9 +9,7 @@
       :routeEditName='routeEditName',
       :storeGetter='storeGetter',
       :storeRemove='storeRemove',
-      :subtitle='subtitle',
       :tableCols='tableCols',
-      :title='title',
       )
 </template>
 
@@ -32,13 +30,11 @@ export default {
       routeEditName: 'LanguagesEdit',
       storeGetter: 'languagesList',
       storeRemove: 'languageRemove',
-      subtitle: 'Languages description',
       tableCols: [
         { label: 'Name', prop: 'name' },
         { label: 'Native name', prop: 'nativeName' },
         { label: 'Code ISO_639-1', prop: 'code' },
       ],
-      title: 'Languages',
     }
   },
 }